// types/api.ts
export interface components {
    schemas: {
// ...
        Textbook: {
            /** @description MongoDB document ObjectID */
            _id?: components["schemas"]["PydanticObjectId"] | null;
            /**
             * Author
             * @description Author of the textbook
             */
            author: string;
            /**
             * Title
             * @description Title of the textbook
             */
            title: string;
            /**
             * Subject
             * @description Subject of the textbook
             */
            subject: string;
            /**
             * Code
             * @description Code of the textbook
             */
            code: string;
            /**
             * Chapters
             * @description List of chapters in the textbook
             */
            chapters: components["schemas"]["Chapter"][];
            /**
             * Cover Url
             * @description URL of the textbook cover image
             */
            cover_url: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at?: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at?: string;
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}